# production_openapi.json
{"openapi":"3.1.0","info":{"title":"Reel Mapper API","version":"0.1.0"},"paths":{"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Read Users Me","description":"Get current user.\n\nReturns the user object from our database, which includes:\n- id: Our internal UUID\n- email: User's email\n- clerk_user_id: The Clerk user ID (for debugging/linking)","operationId":"read_users_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/save-events/":{"post":{"tags":["save-events"],"summary":"Create Save Event","description":"Create new save event and enqueue extraction job.","operationId":"create_save_event_api_v1_save_events__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveEventCreate"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveEventRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/home":{"get":{"tags":["data"],"summary":"Get Home Data","description":"Get homepage data: User's lists and unsorted restaurants.","operationId":"get_home_data_api_v1_home_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HomeResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/restaurants/{restaurant_id}":{"get":{"tags":["data"],"summary":"Get Restaurant","operationId":"get_restaurant_api_v1_restaurants__restaurant_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"string","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestaurantRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/favorites":{"get":{"tags":["data"],"summary":"Get Favorites","operationId":"get_favorites_api_v1_favorites_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoritesListResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/favorites/{restaurant_id}":{"post":{"tags":["data"],"summary":"Toggle Favorite","operationId":"toggle_favorite_api_v1_favorites__restaurant_id__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"string","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoriteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/visited":{"get":{"tags":["data"],"summary":"Get Visited","operationId":"get_visited_api_v1_visited_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitedListResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/visited/{restaurant_id}":{"post":{"tags":["data"],"summary":"Toggle Visited","operationId":"toggle_visited_api_v1_visited__restaurant_id__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"string","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/lists/":{"post":{"tags":["lists"],"summary":"Create List","description":"Create a new list.","operationId":"create_list_api_v1_lists__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/lists/{list_id}/restaurants":{"post":{"tags":["lists"],"summary":"Add Restaurant To List","operationId":"add_restaurant_to_list_api_v1_lists__list_id__restaurants_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"list_id","in":"path","required":true,"schema":{"type":"string","title":"List Id"}},{"name":"restaurant_id","in":"query","required":true,"schema":{"type":"string","title":"Restaurant Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRestaurantRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/lists/{ist_id}":{"delete":{"tags":["lists"],"summary":"Delete List","description":"Delete a list. Items in the list are moved to 'Unsorted' (list_id=None).","operationId":"delete_list_api_v1_lists__ist_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"list_id","in":"query","required":true,"schema":{"type":"string","title":"List Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Delete List Api V1 Lists  Ist Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/lists/{list_id}/restaurants/{restaurant_id}":{"delete":{"tags":["lists"],"summary":"Remove Restaurant From List","description":"Remove a restaurant from a list (move to Unsorted).","operationId":"remove_restaurant_from_list_api_v1_lists__list_id__restaurants__restaurant_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"list_id","in":"path","required":true,"schema":{"type":"string","title":"List Id"}},{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"string","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Remove Restaurant From List Api V1 Lists  List Id  Restaurants  Restaurant Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user-restaurants/{id}":{"delete":{"tags":["user-restaurants"],"summary":"Delete User Restaurant","description":"Delete a saved restaurant.","operationId":"delete_user_restaurant_api_v1_user_restaurants__id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user-restaurants/restaurant/{restaurant_id}":{"delete":{"tags":["user-restaurants"],"summary":"Delete User Restaurant By Rid","description":"Delete a saved restaurant by restaurant_id.","operationId":"delete_user_restaurant_by_rid_api_v1_user_restaurants_restaurant__restaurant_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"string","title":"Restaurant Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"FavoriteResponse":{"properties":{"is_favorite":{"type":"boolean","title":"Is Favorite"}},"type":"object","required":["is_favorite"],"title":"FavoriteResponse"},"FavoritesListResponse":{"properties":{"restaurant_ids":{"items":{"type":"string","format":"uuid"},"type":"array","title":"Restaurant Ids"}},"type":"object","required":["restaurant_ids"],"title":"FavoritesListResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HomeResponse":{"properties":{"lists":{"items":{"$ref":"#/components/schemas/ListRead"},"type":"array","title":"Lists"},"unsorted_restaurants":{"items":{"$ref":"#/components/schemas/UserRestaurantRead"},"type":"array","title":"Unsorted Restaurants"}},"type":"object","required":["lists","unsorted_restaurants"],"title":"HomeResponse"},"ListCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"ListCreate"},"ListRead":{"properties":{"name":{"type":"string","title":"Name"},"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name","id","created_at"],"title":"ListRead"},"RestaurantRead":{"properties":{"name":{"type":"string","title":"Name"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"city":{"type":"string","title":"City"},"price_range":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Price Range"},"google_place_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Google Place Id"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["name","latitude","longitude","city","id"],"title":"RestaurantRead"},"SaveEventCreate":{"properties":{"source_url":{"type":"string","maxLength":2083,"minLength":1,"format":"uri","title":"Source Url"},"raw_caption":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Caption"},"target_list_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Target List Id"}},"type":"object","required":["source_url"],"title":"SaveEventCreate"},"SaveEventRead":{"properties":{"source_url":{"type":"string","maxLength":2083,"minLength":1,"format":"uri","title":"Source Url"},"raw_caption":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Caption"},"target_list_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Target List Id"},"id":{"type":"string","format":"uuid","title":"Id"},"status":{"type":"string","title":"Status"}},"type":"object","required":["source_url","id","status"],"title":"SaveEventRead"},"UserRead":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"id":{"type":"string","format":"uuid","title":"Id"},"clerk_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Clerk User Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},"type":"object","required":["email","id"],"title":"UserRead"},"UserRestaurantRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"restaurant":{"$ref":"#/components/schemas/RestaurantRead"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","restaurant","created_at"],"title":"UserRestaurantRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VisitedListResponse":{"properties":{"restaurant_ids":{"items":{"type":"string","format":"uuid"},"type":"array","title":"Restaurant Ids"}},"type":"object","required":["restaurant_ids"],"title":"VisitedListResponse"},"VisitedResponse":{"properties":{"is_visited":{"type":"boolean","title":"Is Visited"}},"type":"object","required":["is_visited"],"title":"VisitedResponse"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}